requirements.txt
transformers==4.17.0
tqdm==4.63.0
tensorboard==2.8.0
datasets==2.0.0
spacy==2.2.4
matplotlib==3.2.2
gitpython==3.1.12
pytorch_lightning==1.1.3
rouge-score==0.0.4
sacrebleu==1.4.3
sentencepiece==0.1.95
fairscale

